requirements.txt
langgraph>=0.1.0,<0.2.0
langchain-groq>=0.1.0,<0.2.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
pydantic>=2.0.0,<3.0.0
chromadb>=0.4.0,<0.5.0
networkx>=3.0,<4.0
pdfplumber>=0.10.0,<0.11.0
pdf2image>=1.17.0,<2.0.0
Pillow>=10.0.0,<11.0.0
pytesseract>=0.3.10,<0.4.0
jinja2>=3.1.0,<4.0.0
weasyprint>=60.0,<61.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
flask==3.0.3
